Switch from upsertion to staging strategy for attachment driver
Most attachments will be delivered as streams, which can only be
consumed once. But there are things we need to know about data before
accepting them, such as the size and hash. But determining these would
consume the stream. So we need to temporarily write new attachments to a
staging area, determine their info, and then commit/reject based on the
hash+size.